+++ src/components/List/List.module.scss
@@ -38,55 +38,7 @@
     }
 
 }
-.optionBlock{
-    background-color: white;
-    border-radius: 8px;
-    padding: 10px 0px;
-    position: absolute;
-    top: 50px;
-    left: 186px;
-    width: 250px;
-    z-index: 1000;
-    &_header{
-       display: flex;
-       padding-left: 70px;
-       p{
-        align-self: center;
-       }
-       h4 {
-        padding: 2px 8px;
-        margin-left: 50px;
 
-        &:hover {
-            background-color: #747f8f7f;
-            
-        }
-    }
-        
-    }
-}
-.parentStandartDiv{
-    display: flex;
-    flex-direction: column;
-    h3{
-        font-size: 12px;
-        padding: 5px;
-        font-weight: 900;
-        cursor:auto;
-    }
-    hr{
-        margin: 15px;
-    }
-}
-.standartBlock{
-    padding: 6px 8px;
-    cursor: pointer;
-    font-size: 14px;
-    
-    &:hover{
-        background-color: #4e5c7331;
-    }
-}
 .addBlock{
     display: flex;
     padding: 12px 7px;
@@ -128,9 +80,7 @@
     }
    
 }   
-.listNone{
-    display: none;
-}
+
 .titleBlock{
     flex: 1;
     align-self: center;
